+++ src/Components/Chart/Chart.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import "./Chart.css"
+function Chart(props) {
+    return (
+        <div className="chart">
+
+        </div>
+    );
+}
+
+export default Chart;
