+++ src/Components/Chart/ChartBar/ChartBar.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import "./ChartBar.css"
+
+function ChartBar(props) {
+    return (
+        <div className="chart-bar"></div>
+    );
+}
+
+export default ChartBar;
